refactor: detach mainCVSection from mainCollectionView
